feat: update instruct eval
